more test suite cleanup
