+++ test/name.js
@@ -17,3 +17,18 @@ runner.test('name-unicode: unicode names and aliases are permitted', function ()
   a.strictEqual(result.两, '2')
   a.strictEqual(result.три, '3')
 })
+
+runner.test('name-alias-mix: one of each', function () {
+  const optionDefinitions = [
+    { name: 'one', alias: 'o' },
+    { name: 'two', alias: 't' },
+    { name: 'three', alias: 'h' },
+    { name: 'four', alias: 'f' }
+  ]
+  const argv = [ '--one', '-t', '--three' ]
+  const result = commandLineArgs(optionDefinitions, { argv })
+  a.strictEqual(result.one, null)
+  a.strictEqual(result.two, null)
+  a.strictEqual(result.three, null)
+  a.strictEqual(result.four, undefined)
+})
