+++ test/output-camel-case.js
