+++ test/output-grouping.js
